+++ styles/global.js
@@ -1,0 +1,102 @@
+import { StyleSheet } from 'react-native';
+
+export const globalStyles = StyleSheet.create({
+    /*Per Calvin University Brand Identity Standards 
+      Fonts: Constantia, Gotham, Century Schoolbook
+      Maroon: #8C2131
+      Gold:   #F3CD00 
+      Can be found at https://calvin.edu/dotAsset/f784aa74-291f-45b1-b45c-d6455663bcb4
+    */
+    container: {
+      flex: 1,
+      backgroundColor: '#fff',
+      fontFamily: 'Century Schoolbook',
+    },
+  
+    screen: {
+      flex: 1,
+      padding: 20,
+      alignItems: 'center',
+    },
+  
+    header: {
+      padding: 20,
+      fontSize: 30,
+      // fontFamily: 'Constantia',
+      fontWeight: 'bold',
+      alignItems: 'center',
+    },
+  
+    button: {
+      width: 100,
+      height: 50,
+      backgroundColor: '#8C2131',
+      borderRadius: 20,
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+  
+    buttonText: {
+      fontSize: 24,
+      // fontFamily: 'Gotham',
+      color: '#F3CD00'
+    },
+  
+    postsWrapper: {
+      paddingTop: 20,
+      alignItems: 'center',
+    },
+    
+    sectionTitle: {
+      fontSize: 24,
+      fontFamily: 'Gotham',
+      fontSize: 24,
+      fontWeight: "bold",
+    },
+    
+    items: {
+      marginTop: 30,
+      alignItems: 'center',
+      paddingHorizontal: '20',
+    },
+  
+    writePostWrapper: {
+      position: 'absolute',
+      paddingHorizontal: 10,
+      bottom: 60,
+      width: '100%',
+      flexDirection: 'row',
+      justifyContent: 'space-between',
+      alignItems: 'center'
+    },
+  
+    input: {
+      paddingVertical: 15,
+      paddingHorizontal: 15,
+      backgroundColor: '#FFF',
+      borderRadius: 60,
+      borderColor: '#8C2131',
+      borderWidth: 1,
+      width: 250,
+    },
+  
+    addWrapper: {
+      width: 60,
+      height: 60,
+      backgroundColor: '#8C2131',
+      borderRadius: 60,
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+  
+    addText: {
+      color: '#F3CD00',
+      fontSize: 30,
+      fontWeight: 'bold',
+    },
+    paragraphs: {
+      fontSize: 16,
+      padding: 20,
+  
+    },  
+  });
